--- frontend/src/agents/events.rs
@@ -1,14 +1,0 @@
-use serde::{Deserialize, Serialize};
-
-#[derive(Clone, Copy, Serialize, Deserialize, Debug)]
-pub enum GlobalEvent {
-    SocketStatus {
-        connected: bool,
-        timeout_secs: Option<i64>,
-    },
-    OpenSharePopup,
-    OpenQuestionPopup,
-    DeletePopup,
-    QuestionCreated(i64),
-    PayForUpgrade,
-}
