+++ frontend/src/agents/mod.rs
@@ -1,5 +1,3 @@
-mod events;
 mod ws_agent;
 
-pub use events::GlobalEvent;
 pub use ws_agent::{SocketInput, WebSocketAgent, WsResponse};
